requirements.txt
setuptools==67.6.1
torch>=2.0.0
numpy~=1.24.2
matplotlib~=3.7.1
seaborn~=0.12.2
prettytable~=3.7.0
tqdm~=4.65.0
scikit-learn~=1.2.2
wandb~=0.14.2
tokenizers==0.13.3
transformers==4.30.1
PyYAML~=6.0
rich==13.3.4
jsonlines~=3.1.0
datasets==2.11.0
convokit~=2.5.3
torchinfo~=1.7.2
huggingface-hub~=0.14.1
